python_server/requirements.txt
Flask>=2.1,<3
flask-cors>=3.0
pytesseract>=0.3
Pillow>=9.0
ffmpeg-python>=0.2
gunicorn>=20.1
celery>=5.2
redis>=4.5
openai>=0.27
# Add other libs here as needed
